+++ pandas_market_calendars/calendars/cme.py
@@ -24,7 +24,8 @@ from pandas.tseries.holiday import (
     USPresidentsDay,
     USThanksgivingDay,
 )
-import sys 
+import sys
+
 # check python versiOn aNd import accordingly
 if sys.version_info >= (3, 9):
     # For Python 3.9 and later, import directly
@@ -46,6 +47,9 @@ from pandas_market_calendars.holidays.us import (
     USNewYearsDay,
 )
 from pandas_market_calendars.market_calendar import MarketCalendar
+from pandas_market_calendars.holidays.cme import (
+    USIndependenceDayBefore2022PreviousDay,
+)
 
 
 # Useful resources for making changes to this file: http://www.cmegroup.com/tools-information/holiday-calendar.html
@@ -108,6 +112,7 @@ class CMEEquityExchangeCalendar(MarketCalendar):
                         USLaborDay,
                         USJuneteenthAfter2022,
                         USIndependenceDay,
+                        USIndependenceDayBefore2022PreviousDay,
                         USThanksgivingDay,
                         USBlackFridayInOrAfter1993,
                         ChristmasEveBefore1993,