+++ pandas_market_calendars/calendars/cme_globex_energy_and_metals.py
@@ -18,7 +18,8 @@ from datetime import time
 from pandas.tseries.holiday import (
     AbstractHolidayCalendar,
 )  # , GoodFriday, USLaborDay, USPresidentsDay, USThanksgivingDay
-import sys 
+import sys
+
 # check python versiOn aNd import accordingly
 if sys.version_info >= (3, 9):
     # For Python 3.9 and later, import directly