+++ pandas_market_calendars/calendars/eurex_fixed_income.py
@@ -6,7 +6,8 @@ from pandas.tseries.holiday import (
     GoodFriday,
     Holiday,
 )
-import sys 
+import sys
+
 # check python versiOn aNd import accordingly
 if sys.version_info >= (3, 9):
     # For Python 3.9 and later, import directly