+++ pandas_market_calendars/calendars/iex.py
@@ -3,7 +3,8 @@ from itertools import chain
 
 from pandas import Timestamp, DatetimeIndex, Timedelta
 from pandas.tseries.holiday import AbstractHolidayCalendar
-import sys 
+import sys
+
 # check python versiOn aNd import accordingly
 if sys.version_info >= (3, 9):
     # For Python 3.9 and later, import directly