+++ pandas_market_calendars/calendars/lse.py
@@ -16,7 +16,8 @@
 from datetime import time
 
 from pandas.tseries.holiday import AbstractHolidayCalendar, EasterMonday, GoodFriday
-import sys 
+import sys
+
 # check python versiOn aNd import accordingly
 if sys.version_info >= (3, 9):
     # For Python 3.9 and later, import directly