+++ pandas_market_calendars/calendars/tase.py
@@ -2,7 +2,8 @@ from datetime import time
 
 from typing import Literal, Union
 from pandas import Timestamp, Timedelta, DatetimeIndex
-import sys 
+import sys
+
 # check python versiOn aNd import accordingly
 if sys.version_info >= (3, 9):
     # For Python 3.9 and later, import directly