+++ pandas_market_calendars/calendars/tsx.py
@@ -10,7 +10,8 @@ from pandas.tseries.holiday import (
     MO,
     weekend_to_monday,
 )
-import sys 
+import sys
+
 # check python versiOn aNd import accordingly
 if sys.version_info >= (3, 9):
     # For Python 3.9 and later, import directly